Total rewrite.
This new version doesn't try to parse client specs.
Instead it just asks the server for all the files in the client workspace
that it knows about, and filters those from the complete list of files.

For the first cut, no attempt is made to skip files which cannot be added
to the depot because of exclusionary client mappings.